api/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-core
azure-search-documents==11.6.0b1
openai
python-dotenv
pydantic
fastapi
azure-cognitiveservices-speech
python-multipart
azure-storage-blob
pydub
